# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for a plant nursery point-of-sale system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      responses:
        "204":
          description: Deleted
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      responses:
        "204":
          description: Deleted
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice (complete a sale)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get an invoice
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
components:
  schemas:
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, createdAt]
    SupplierInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
      required: [name]
    Product:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
          nullable: true
        barcode:
          type: string
        categoryId:
          type: integer
          nullable: true
        categoryNameAr:
          type: string
          nullable: true
        supplierId:
          type: integer
          nullable: true
        purchasePrice:
          type: number
        salePrice:
          type: number
        quantity:
          type: integer
        lowStockThreshold:
          type: integer
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
      required:
        - id
        - nameAr
        - barcode
        - purchasePrice
        - salePrice
        - quantity
        - lowStockThreshold
        - status
        - createdAt
    ProductInput:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
          nullable: true
        barcode:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        supplierId:
          type: integer
          nullable: true
        purchasePrice:
          type: number
        salePrice:
          type: number
        quantity:
          type: integer
        lowStockThreshold:
          type: integer
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
      required: [nameAr, purchasePrice, salePrice, quantity]
...
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productNameAr:
          type: string
        quantity:
          type: integer
        price:
          type: number
        discount:
          type: number
        notes:
          type: string
          nullable: true
      required: [id, productId, productNameAr, quantity, price, discount]
    InvoiceItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        discount:
          type: number
        notes:
          type: string
          nullable: true
      required: [productId, quantity, price]
    Invoice:
      type: object
      properties:
        id:
          type: integer
        number:
          type: string
        customerId:
          type: integer
          nullable: true
        customerName:
          type: string
          nullable: true
        employeeId:
          type: integer
          nullable: true
        employeeName:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cash, visa, cliq, bank, split]
        status:
          type: string
          enum: [completed, refunded]
        createdAt:
          type: string
      required:
        - id
        - number
        - items
        - subtotal
        - discount
        - tax
        - total
        - paymentMethod
        - status
        - createdAt
    InvoiceInput:
      type: object
      properties:
        customerId:
          type: integer
          nullable: true
        employeeId:
          type: integer
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
        discount:
          type: number
        tax:
          type: number
        paymentMethod:
          type: string
          enum: [cash, visa, cliq, bank, split]
      required: [items, paymentMethod]
...