# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for a plant nursery point-of-sale system
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface Product {
  id: number;
  nameAr: string;
  nameEn?: string | null;
  barcode: string;
  categoryId?: number | null;
  categoryNameAr?: string | null;
  supplierId?: number | null;
  purchasePrice: number;
  salePrice: number;
  quantity: number;
  lowStockThreshold: number;
  description?: string | null;
  imageUrl?: string | null;
  status: ProductStatus;
  createdAt: string;
}