# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for a plant nursery point-of-sale system
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputStatus } from './productInputStatus';

export interface ProductInput {
  nameAr: string;
  nameEn?: string | null;
  barcode?: string | null;
  categoryId?: number | null;
  supplierId?: number | null;
  purchasePrice: number;
  salePrice: number;
  quantity: number;
  lowStockThreshold?: number;
  description?: string | null;
  imageUrl?: string | null;
  status?: ProductInputStatus;
}